backend/requirements.txt
asgiref==3.10.0
cffi==2.0.0
colorama==0.4.6
cryptography==46.0.3
Django==5.2.7
django-cors-headers==4.9.0
django-extensions==4.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
MarkupSafe==3.0.3
pillow==12.0.0
psycopg2-binary==2.9.11
pycparser==2.23
PyJWT==2.10.1
pyOpenSSL==25.3.0
qrcode==8.2
sqlparse==0.5.3
tzdata==2025.2
Werkzeug==3.1.3
gunicorn
dj-database-url
